# projects/counter-api/api/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "github.com/JamesArthurHolland/ivanti/profile-api",
    "description": "Boilerplate code generated by **sqlc-grpc**. Modify _proto/*.proto_ files then run `buf generate` to change the services interface.",
    "version": "1.0",
    "contact": {
      "name": "sqlc-grpc",
      "url": "https://github.com/walterwanderley/sqlc-grpc"
    }
  },
  "tags": [
    {
      "name": "CounterService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/counter": {
      "put": {
        "operationId": "CounterService_InsertCounter",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1Counter"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1InsertCounterRequest"
            }
          }
        ],
        "tags": [
          "CounterService"
        ]
      }
    },
    "/fetch-counter/{id}": {
      "get": {
        "operationId": "CounterService_FetchCounter",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1FetchCounterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "CounterService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Counter": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "count": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1FetchCounterResponse": {
      "type": "object",
      "properties": {
        "value": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1InsertCounterRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "v1InsertCounterResponse": {
      "type": "object",
      "properties": {
        "counter": {
          "$ref": "#/definitions/v1Counter"
        }
      }
    }
  }
}